requirements.txt
requests
werkzeug
Pillow
flask
flask-login
flask-sqlalchemy